Rename former c2p ocm plugins
