+++ cmd/ocm/oscal2policy/c2p-config.yaml
