+++ cmd/ocm/oscal2policy/options/options.go
@@ -34,9 +34,9 @@ func NewOptions() *Options {
 }
 
 func (o *Options) AddFlags(fs *pflag.FlagSet) {
-	fs.StringVar(&o.C2PCRPath, "c2pcr", "", "path to c2p CR")
+	fs.StringVarP(&o.C2PCRPath, "config", "c", "", "path to c2p config file")
 	fs.StringVar(&o.TempDirPath, "temp-dir", "", "path to temp directory")
-	fs.StringVar(&o.OutputDir, "out", ".", "path to a directory for output manifest files of generated OCM Policy manifests")
+	fs.StringVarP(&o.OutputDir, "out", "o", ".", "path to a directory for output manifest files of generated OCM Policy manifests")
 	fs.StringVar(&o.OutputDirForPolicyGenerator, "out-for-policy-generator", "", "path to a directory for output files for policy generator to generate OCM Policy manifests (default: system temporary directory or directory specified by --temp-dir)")
 }
 
@@ -46,7 +46,7 @@ func (o *Options) Complete() error {
 
 func (o *Options) Validate() error {
 	if o.C2PCRPath == "" {
-		return errors.New("--c2pcr is required")
+		return errors.New("-c or --config is required")
 	}
 	return nil
 }
