+++ cmd/ocm/result2oscal/c2p-config.yaml
