+++ pkg/ocm/oscal2policy.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package composer
+package ocm
 
 import (
 	"fmt"
